% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applySysConfig}
\alias{applySysConfig}
\title{Apply a system configuration to a compatible model}
\usage{
applySysConfig(sysConfig, modelSystem)
}
\arguments{
\item{sysConfig}{a CLR object that implements ISystemConfiguration}

\item{modelSystem}{the modelling system to configure}
}
\description{
Apply a system configuration to a compatible model
}

