% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objectives.r
\name{asDataFrame}
\alias{asDataFrame}
\title{Convert the content of a collection of optim results as a data frame}
\usage{
asDataFrame(mhObject)
}
\arguments{
\item{mhObject}{An object that can be referenced as an IEnumerable<IObjectiveScores>. Further types may be supported later.}
}
\value{
a data frame
}
\description{
Convert the content of a collection of optim results (such as the output of an optimisation process) as a data frame
}

