% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.r
\name{boundValues}
\alias{boundValues}
\title{min/max bound numeric values}
\usage{
boundValues(x, lim = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{lim}{a num vector of the min/max limits to apply, for instance c(0, 1)}
}
\value{
a numeric vector
}
\description{
min/max bound numeric values
}

