% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boundValuesDf}
\alias{boundValuesDf}
\title{min/max bound a column in a data frame}
\usage{
boundValuesDf(x, colname, lim = c(0, 1))
}
\arguments{
\item{x}{a data frame}

\item{colname}{a character vector, name of the column to bound}

\item{lim}{a num vector of the min/max limits to apply, for instance c(0, 1)}
}
\value{
a data frame
}
\description{
min/max bound a column in a data frame
}

