% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildParamSet}
\alias{buildParamSet}
\title{Builds a parameter set from a line in a data frame.}
\usage{
buildParamSet(paramSet, dataFrame, rownum = 1)
}
\arguments{
\item{paramSet}{A parameter set template, an object that implements IHyperCubeSetBounds<double>}

\item{dataFrame}{a data frame with at least column names including all the names of the template parameter set.}

\item{rownum}{The row index to use; defaults to 1}
}
\value{
A new parameter set, cloned from the template and with updated values
}
\description{
Builds a parameter set from a line in a data frame, typically a line from a calibration log.
}

