% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coeffVariationTermination}
\alias{coeffVariationTermination}
\title{Create an termination criterion based on a coefficient of variation of the parameter sets}
\usage{
coeffVariationTermination(sysConfigType = hyperCubeType(),
  cvThreshold = 0.025, maxHours = 0.1)
}
\arguments{
\item{sysConfigType}{The type of the system configuration to use for Generics.}

\item{cvThreshold}{CV of the parameter sets, below which the process has converged. It is applied to a subset of the population of points (the "better" half of it)}

\item{maxHours}{the maximum wall time runtime for the optimisation}
}
\value{
a CLR termination criterion.
}
\description{
Create an termination criterion based on a coefficient of variation of the parameter sets
}

