% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copyMhData}
\alias{copyMhData}
\title{A copy constructor for mhData objects}
\usage{
copyMhData(src, newData)
}
\arguments{
\item{src}{the object to use as a source for the object member values.}

\item{newData}{a data frame}
}
\value{
mhData object
}
\description{
A copy constructor for mhData objects
}

