% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimizers.r
\name{createSceOptim}
\alias{createSceOptim}
\title{Build an SCE optimizer}
\usage{
createSceOptim(objective, parameterizer, sysConfigType = NULL,
  terminationCriterion = maxWallTimeTermination(), sceParameters = NULL)
}
\arguments{
\item{objective}{an objective calculator}

\item{parameterizer}{a type that is implementing IHyperCube<double>}

\item{sysConfigType}{The type of the system configuration to use for Generics... type of parameterizer if this is the default value NULL}

\item{terminationCriterion}{An object that can be passed to SCE for testing the completion of the algorithm.}

\item{sceParameters}{An object of CLR type SceParameters, or NULL for defaults.}
}
\value{
a Shuffled Complex Evolution optimizer
}
\description{
Build an SCE optimizer
}

