% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizers.r
\name{createSceParameters}
\alias{createSceParameters}
\title{Builds a configuration for SCE optimizer}
\usage{
createSceParameters(alpha = 1, beta = 9, contractRatio = 0.5,
  reflectRatio = -1, m = 9, p = 5, pmin = 5, q = 5,
  trapezDensity = 2, numShuffles = -1)
}
\arguments{
\item{alpha}{Number of geometrical transformation for each subcomplex}

\item{beta}{Number of evolution steps taken by each sub-complex before shuffling occurs}

\item{contractRatio}{The homothetic ratio used in the contraction phase of the complex evolution: default 0.5}

\item{reflectRatio}{The homothetic ratio used in the reflection phase of the complex evolution: default -1.0}

\item{m}{Number of points per complex}

\item{p}{Number of complexes}

\item{pmin}{Minimum number of complexes (populations of points)}

\item{q}{Number of points per SUB-complex}

\item{trapezDensity}{a coefficient between 0 and 2 (though between 1 and 2 only makes practical sense) defining the probability for assigning points to subcomplexes. Defaults to 2}

\item{numShuffles}{maximum number of shuffles. -1 by default means that the termination criterion ignores it.}
}
\value{
a Shuffled Complex Evolution parameters object
}
\description{
Builds a configuration for SCE optimizer
}

