% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizers.r
\name{createSceParamForDimension}
\alias{createSceParamForDimension}
\title{Builds an SCE parameterization for a problem with N degrees of freedom}
\usage{
createSceParamForDimension(n)
}
\arguments{
\item{n}{degrees of freedom}
}
\value{
a Shuffled Complex Evolution parameters object
}
\description{
Builds an SCE parameterization for a problem with N degrees of freedom. This is based on rule of thumbs published in (REF to add)
}

