% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-database.R
\name{db_fetch_chunks}
\alias{db_fetch_chunks}
\title{Fetch data from query in chunks}
\usage{
db_fetch_chunks(res, progress = TRUE, chunk_size = 0.01)
}
\arguments{
\item{res}{Object returned by \code{DBI::dbSendQuery}}

\item{progress}{Show progress in console?}

\item{chunk_size}{Number of rows to be fethced in each loop. Number smaller
than 1 are itertreted as fractions of total number of affected rows.}
}
\value{
data.table
}
\description{
Fetch data from query in chunks
}
