% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.r
\name{dotOnLoad}
\alias{.onLoad}
\alias{dotOnLoad}
\title{Function called when loading the package with 'library'.}
\usage{
.onLoad(libname = "~/R", pkgname = "mh")
}
\arguments{
\item{libname}{the path to the library from which the package is loaded}

\item{pkgname}{the name of the package.}
}
\description{
Function called when loading the package with 'library'.
}

