% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-misc.R
\name{encode_chr_vars<-}
\alias{encode_chr_vars<-}
\alias{encode_chr_vars<-.data.frame}
\alias{encode_chr_vars<-.data.table}
\title{Set encoding for entire columns in data.frame or data.table}
\usage{
encode_chr_vars(x) <- value

\method{encode_chr_vars}{data.frame}(x) <- value

\method{encode_chr_vars}{data.table}(x) <- value
}
\arguments{
\item{x}{A data.frame or data.table object.}

\item{value}{Encoding to set.}
}
\description{
Wrapper around \code{\link{Encoding}}. Sets all character columns to given encoding.
more replacements and uses camel case. Caveat: Factors are ignored.
}
\examples{
 xx <- data.frame(a = LETTERS, b = letters, c = 1:26, stringsAsFactors = F)
}
