% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{format_int}
\alias{format_int}
\alias{format_dec}
\title{Quick formatting of numbers}
\usage{
format_int(x, width = 2, flag = "0")

format_dec(x, digits = 2)
}
\description{
Wrapper for \code{\link{formatC}}.
}
