% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLoggerContent}
\alias{getLoggerContent}
\title{Gets the content of a calibration log as a data frame}
\usage{
getLoggerContent(calibLogger)
}
\arguments{
\item{calibLogger}{an object implementing ILoggerMh. May be a missing argument, in which case a defaut in-memory logger is created.}
}
\value{
a data frame
}
\description{
Gets the content of a calibration log as a data frame
}

