% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameters.r
\name{getMaxValue}
\alias{getMaxValue}
\title{Get the maximum value(s) of a hypercube parameter set}
\usage{
getMaxValue(paramSet, varName)
}
\arguments{
\item{paramSet}{An object that implements at least IHyperCube<double>}

\item{varName}{character vector, name(s) of the variable(s) to get}
}
\value{
numeric vector, value(s) of the maximum value(s) of the variable(s).
}
\description{
Get the maximum value(s) of a hypercube parameter set
}

