% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objectives.r
\name{getScore}
\alias{getScore}
\title{Calculate the score of a given system configuration}
\usage{
getScore(objectiveEvaluator, pSet)
}
\arguments{
\item{objectiveEvaluator}{an objective calculator (implements at least IObjectiveEvaluator)}

\item{pSet}{the system configuration that the calculator accept to calculate a goodness of fit. (implements at least ISystemConfiguration)}
}
\value{
a list of length two: a named numeric vector, with one or more objective values, and a representation of the parameters (possibly as a data frame)
}
\description{
Calculate the score of a given system configuration
}

