% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.r
\name{getValue}
\alias{getValue}
\title{Get the value(s) of a hypercube parameter set}
\usage{
getValue(paramSet, varName)
}
\arguments{
\item{paramSet}{An object that implements at least IHyperCube<double>}

\item{varName}{character vector, name(s) of the variable(s) to get}
}
\value{
numeric vector, value(s) of the variable(s) value(s).
}
\description{
Get the value(s) of a hypercube parameter set
}

