% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hyperCubeSetBoundsType}
\alias{hyperCubeSetBoundsType}
\title{Gets the CLR type for IHyperCubeSetBounds<double>}
\usage{
hyperCubeSetBoundsType()
}
\value{
a System.Type
}
\description{
Gets the CLR type for IHyperCubeSetBounds<double>. This function is sometimes needed to pass arguments to some functions requiring an explicit type of system configuration, e.g. createSceOptim.
}

