% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-misc.R
\name{make_names}
\alias{make_names}
\alias{make_names.default}
\alias{make_names.data.frame}
\title{Prettify variable names}
\usage{
make_names(x, ...)

\method{make_names}{default}(x)

\method{make_names}{data.frame}(x)
}
\arguments{
\item{x}{A character vector or data frame}
}
\description{
Similar to \code{\link{make.names}} but does
more replacements and uses camel case.
}
\examples{
 library(dplyr)

 # workhorse: character vector
 "12var_äüß//__stuff" \%>\% make_names

 # dataframe: make pretty colnames
 iris \%>\% make_names \%>\% names

}
