% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizers.r
\name{marginalImprovementTermination}
\alias{marginalImprovementTermination}
\title{Create an termination criterion based on the rate of marginal fitness improvement}
\usage{
marginalImprovementTermination(sysConfigType = hyperCubeType(),
  tolerance = 1e-06, cutoffNoImprovement = 10, maxHours = 0.1)
}
\arguments{
\item{sysConfigType}{The type of the system configuration to use for Generics.}

\item{tolerance}{the increment in the objective below which the improvement is considered negligible}

\item{cutoffNoImprovement}{the maximum number of successive times the algorithm fails to improve the objective function.}

\item{maxHours}{the maximum wall time runtime for the optimisation}
}
\value{
a CLR termination criterion.
}
\description{
Create an termination criterion based on the rate of marginal fitness improvement
}

