% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plotParamEvolution}
\alias{plotParamEvolution}
\title{Plot the value of a parameter along the optimisation process}
\usage{
plotParamEvolution(logMh, paramName, objLims = NULL)
}
\arguments{
\item{logMh}{an mhData object}

\item{paramName}{the exact name of one of the model parameters}

\item{objLims}{optional bounds to apply to the objective function value prior to plotting (e.g. c(0,1) for Nash-Sutcliffe efficiency)}
}
\value{
a ggplot object
}
\description{
Plot the value of a parameter along the optimisation process.
The color scale is the objective score. Useful to check the behavior of the optimisation process.
}
\examples{
\dontrun{
logFileName <- 'F:/path/to/419016_IhacresClassic_19390101_19481231.csv'
logSce <- loadMhLog(logFileName)
logMh <- new("mhData",
   data= logSce,
   fitness = "NSE.logbias",
   messages = "Message",
   categories = "Category")

geomOps <- subsetByMessage(logMh)
d <- plotParamEvolution(geomOps, 'Tq', objLims=c(0,1))
d
}
}

