% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plotParamEvolutionMsg}
\alias{plotParamEvolutionMsg}
\title{Plot the value of a parameter along the optimisation process}
\usage{
plotParamEvolutionMsg(logMh, paramName)
}
\arguments{
\item{logMh}{an mhData object}

\item{paramName}{the exact name of one of the model parameters}
}
\value{
a ggplot object
}
\description{
Plot the value of a parameter along the optimisation process.
The color scale is the message associated with the point, for instance a geometric transformation.
An example message is 'Contracted point in the subcomplex' for the Shuffled complex algorithm.
Useful to check the behavior of the optimisation process.
}
\examples{
\dontrun{
logFileName <- 'F:/path/to/419016_IhacresClassic_19390101_19481231.csv'
logSce <- loadMhLog(logFileName)
logMh <- new("mhData",
   data= logSce,
   fitness = "NSE.logbias",
   messages = "Message",
   categories = "Category")

geomOps <- subsetByMessage(logMh)
d <- plotParamEvolutionMsg(geomOps, 'Tq')
d
# If there is overplotting, one can use facets to have a clearer view of the optimiser behavior
d + facet_wrap( as.formula(paste("~", geomOps@messages, sep=' ')) )
}
}

