% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.r
\name{processLogger}
\alias{processLogger}
\title{Extract information from an optimisation logger}
\usage{
processLogger(optimizer, fitness = "NSE", messages = "Message",
  categories = "Category")
}
\arguments{
\item{optimizer}{the instance of the optimizer}

\item{fitness}{the name of the fitness score that was used for the optimization, e.g. 'NSE'}

\item{messages}{}

\item{categories}{}
}
\value{
an object for graphing through the 'mh' package functions
}
\description{
Extract information from an optimisation logger. This retrieves data and transforms it into a format more amenable to visualization.
}

