% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pSetAsDataFrame}
\alias{pSetAsDataFrame}
\title{Convert a hypercube to a simple data frame representation}
\usage{
pSetAsDataFrame(paramSet)
}
\arguments{
\item{paramSet}{a CLR object that implements IHyperCube}
}
\value{
a data frame with column names Name,Value,Min,Max,Description
}
\description{
Convert a hypercube to a simple data frame representation
}

