% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.r
\name{rollingMax}
\alias{rollingMax}
\title{Rolling max function}
\usage{
rollingMax(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the rolling maximum over the input vector
}
\description{
Rolling max function
}

