% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.r
\name{setHyperCube}
\alias{setHyperCube}
\title{Configure a hypercube as per a data frame specification}
\usage{
setHyperCube(paramSet, specsFrame, setBounds = FALSE)
}
\arguments{
\item{paramSet}{An object that implements IHyperCubeSetBounds<double>}

\item{specsFrame}{a data frame with column names Name,Value,Min,Max. Min and Max are optional if setBounds is FALSE}

\item{setBounds}{if TRUE, use the Min and Max columns to modify the bounds of the hypercube. Defaults to FALSE.}
}
\description{
Configure a hypercube as per a data frame specification
}

