% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.r
\name{setMinValue}
\alias{setMinValue}
\title{Change the minimum bound(s) of a hypercube parameter set}
\usage{
setMinValue(paramSet, varName, value)
}
\arguments{
\item{paramSet}{An object that implements IHyperCubeSetBounds<double>}

\item{varName}{character vector, name(s) of the variable(s) to change}

\item{value}{numeric vector, value(s) to set the variable(s) minimum bound(s) to.}
}
\description{
Change the minimum bound(s) of a hypercube parameter set
}

