% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameters.r
\name{setValue}
\alias{setValue}
\title{Change the value(s) of a hypercube parameter set}
\usage{
setValue(paramSet, varName, value)
}
\arguments{
\item{paramSet}{An object that implements IHyperCubeSetBounds<double>}

\item{varName}{character vector, name(s) of the variable(s) to change}

\item{value}{numeric vector, value(s) to set the variable(s) value(s) to.}
}
\description{
Change the value(s) of a hypercube parameter set
}

