% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subsetByCategory}
\alias{subsetByCategory}
\title{Subset an mhData object based on string pattern matching}
\usage{
subsetByCategory(logMh, pattern = "Initial.*|Shuffling.*")
}
\arguments{
\item{logMh}{an mhData object}

\item{pattern}{a pattern suitable for use by \code{\link{str_detect}}, for instance 'Initial.*|Reflec.*|Contrac.*|Add.*'}
}
\value{
an mhData object
}
\description{
Subset an mhData object based on string pattern matching on the column for 'categories'
}

