% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subsetByPattern}
\alias{subsetByPattern}
\title{Subset a data frame based on string pattern matching}
\usage{
subsetByPattern(x, colname, pattern)
}
\arguments{
\item{x}{the data frame}

\item{colname}{column name}

\item{pattern}{a pattern suitable for use by \code{\link{str_detect}}, for instance 'Initial.*|Reflec.*|Contrac.*|Add.*'}
}
\value{
a data frame. Unnecessary levels have been dropped from factor columns.
}
\description{
Subset a data frame based on string pattern matching on one of its character columns. Uses the function \code{\link{str_detect}}
}

