% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hyperCubeType}
\alias{hyperCubeType}
\title{Gets the CLR type for IHyperCube<double>}
\usage{
hyperCubeType()
}
\value{
a System.Type
}
\description{
Gets the CLR type for IHyperCube<double>. This function is sometimes needed to pass arguments to some functions requiring an explicit type of system configuration, e.g. createSceOptim.
}

