% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadMhLog}
\alias{loadMhLog}
\title{Load a CSV log file of an optimisation}
\usage{
loadMhLog(fn)
}
\arguments{
\item{fn}{the file name of the CSV}
}
\value{
a data frame, as loaded with read.csv, and an added column 'PointNumber'
}
\description{
Load a CSV log file of an optimisation
}

