% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimizers.r
\name{maxSceShuffleTermination}
\alias{maxSceShuffleTermination}
\title{Create an termination criterion using SCE's maximum number of shuffles}
\usage{
maxSceShuffleTermination(sysConfigType = hyperCubeType())
}
\arguments{
\item{sysConfigType}{The type of the system configuration to use for Generics.}
}
\value{
a CLR termination criterion.
}
\description{
Create an termination criterion using SCE's maximum number of shuffles
}

