% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizers.r
\name{maxWallTimeTermination}
\alias{maxWallTimeTermination}
\title{Create an termination criterion with a max wall time}
\usage{
maxWallTimeTermination(sysConfigType = hyperCubeType(), maxHours = 0.1)
}
\arguments{
\item{sysConfigType}{The type of the system configuration to use for Generics.}

\item{maxHours}{the maximum wall time runtime for the optimisation}
}
\value{
a CLR termination criterion.
}
\description{
Create an termination criterion with a max wall time
}

