% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mh-package.r
\docType{package}
\name{mh-package}
\alias{mh}
\alias{mh-package}
\title{An R programming interface and utilities to access the metaheuristics framework}
\description{
An R programming interface and utilities to access the metaheuristics framework
}
\details{
\tabular{ll}{
Package: \tab mh\cr
Type: \tab Package\cr
Version: \tab 0.4-4\cr
Date: \tab 2015-03-27\cr
Release Notes: \tab Update dependencies due to a change in parameter logging precision to double,in CSIRO.Metaheuristics \cr
License: \tab LGPL 2.1\cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{CLR}
\keyword{Mono}
\keyword{metaheuristics}

