% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{mh-package}
\alias{mh}
\alias{mh-package}
\title{An R programming interface and utilities to access the metaheuristics framework}
\description{
An R programming interface and utilities to access the metaheuristics framework
}
\details{
\tabular{ll}{
Package: \tab mh\cr
Type: \tab Package\cr
Version: \tab 0.4-3\cr
Date: \tab 2015-02-09\cr
Release Notes: \tab Non parallel Calibration with reentrant call to an R function for the objective evaluation from SCE works on Linux \cr
License: \tab LGPL 2.1\cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{CLR}
\keyword{Mono}
\keyword{metaheuristics}

