% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mh-package.r
\docType{package}
\name{mh-package}
\alias{mh}
\alias{mh-package}
\title{An R programming interface and utilities to access the metaheuristics framework}
\description{
An R programming interface and utilities to access the metaheuristics framework
}
\details{
\tabular{ll}{
Package: \tab mh\cr
Type: \tab Package\cr
Version: \tab 0.4-6\cr
Date: \tab 2015-05-07\cr
Release Notes: \tab Fix issue building on Ubuntu and R 3.2 \cr
License: \tab LGPL 3\cr
}

\tabular{lll}{
Version \tab Date \tab Notes \cr
0.4-5: \tab 2015-04-04 \tab Update to use CSIRO.Metaheuristics 0.4.9 - use parallel.foreach to work around Mono 3.12.1 \cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{CLR}
\keyword{Mono}
\keyword{metaheuristics}

