% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.r
\name{mkOptimLog}
\alias{mkOptimLog}
\title{Extract information from an optimisation logger}
\usage{
mkOptimLog(logDataFrame, fitness = "NSE", messages = "Message",
  categories = "Category")
}
\arguments{
\item{logDataFrame}{a data frame, extracted from an optimization logger with mh::getLoggerContent}

\item{fitness}{the name of the fitness score that was used for the optimization, e.g. 'NSE'}

\item{messages}{}

\item{categories}{}
}
\value{
an object for graphing through the 'mh' package functions
}
\description{
Extract information from an optimisation logger. This retrieves data and transforms it into a format more amenable to visualization.
}

