% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.r
\name{pSetAsDataFrame}
\alias{pSetAsDataFrame}
\title{Convert a hypercube to a simple data frame representation}
\usage{
pSetAsDataFrame(paramSet)
}
\arguments{
\item{paramSet}{a CLR object that implements IHyperCube}
}
\value{
a data frame with column names Name,Value,Min,Max,Description
}
\description{
Convert a hypercube to a simple data frame representation
}

