% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setLogger}
\alias{setLogger}
\title{Sets a logger onto an optimizer}
\usage{
setLogger(optimizer, calibLogger)
}
\arguments{
\item{optimizer}{an IOptimizer}

\item{calibLogger}{an object implementing ILoggerMh. May be a missing argument, in which case a defaut in-memory logger is created.}
}
\description{
Sets a logger onto an optimizer
}

